Speed up EncryptRTP and make it append
new:
BenchmarkEncryptRTP-8          	  300000    5181 ns/op
BenchmarkEncryptRTPInPlace-8   	  300000    4480 ns/op

old:
BenchmarkEncryptRTP-8          	  200000    6378 ns/op
BenchmarkEncryptRTPInPlace-8   	  300000    4448 ns/op

The old code would:

1. `allocateIfMismatch` performs `bytes.Equal` to check if the dst/src
are the same. It's not needed, it's technically wrong, and it could
have been done faster with pointers anyway.
    ex: `&dst[0] == &src[0]`

2. `allocateIfMismatch` would always copy the payload, even if we're
just about to overwrite it with encrypted data.

3. `allocateIfMismatch` will make a byte slice of the size/cap of the
payload slice, but won't include extra room for the auth tag (10 bytes)
we're expecting to append. This means that we will always cause an
extra realloc/copy.

The unit test to check if the encryption was done in place was broken.
The ciphertext was being written to the dst, but then append was used
to write the auth tag. This causes a reallocation unless the dst is
allocated with 10 extra bytes of overhead. So encrypting in place would
often not work.