Constant time comparison for auth verification
It's a subtle change. Get it?

Timing attacks rely on measuring the amount of time it takes for an
operation to complete. The idea is that programs will break out early on
error, and sometimes it can be abused.

In this case, an attacker could measure the time it takes to verify an
256 auth tags, each one starting with a different byte. `bytes.Equal`
will immediately return false if the first byte is wrong, otherwise
check the next byte and so on. This means that just one of the auth tags
will take longer than the rest, because the first byte will match.
You can repeat the process until you figure out the correct auth tag.