Fix out-of-bounds slice access in DecryptRTCP
Slice was truncated to exclude the authentication portion, but was
then accessed beyond the truncated portion to read the index. Test
did not catch it because slice backing memory happened to be valid.

Replaced unsetting of 'encrypted' bit in memory with a bitmask on
index, as that is all it was needed for. With only the above fix,
tests would now fail. Both changes are necessary for tests to pass.