Add support for AES128_CM_HMAC_SHA1_32
Adding support for SRTP_AES128_CM_HMAC_SHA1_32 as described in RFC 5764.
This necessitated some light refactoring of how `srtpCipher`s interact
with `ProtectionProfile`: the handling of this ciphersuite is identical
to handling for SRTP_AES128_CM_HMAC_SHA1_80, but the parameters are
different. This change assigns a profile to a cipher when it is created,
and all profile parameters are delegated to the profile (rather than the
previous setup, which split the values of profile parameters across the
profile and the cipher). This change also splits `authTagLen` into
`rtpAuthTagLen` and `rtcpAuthTagLen`, since they can have different
values depending on the profile in use.