Split RTCP compound packet before forwarding
When a compound RTCP packet is received, it is forwarded as is to the
read streams. Downstream components have to check for the SSRC in the
packets matching the track(s) it is handling.

This leads to situations where downstream components could see duplicate
RTCP reports. Happens when there is a downstream handler for all SSRCs.
It receives the compound packet, unmarshals it and invokes the handlers.
As it is fielding all SSRCs, it will get the same compound packet `n`
times and invoke the handlers `n` times.

This PR splits up the compound packet and forwards individual packets to
avoid the end handlers from seeing duplicates. API wise, it is
compatible as it still emits an encoded/marshaled packet. But, this does
add a marshaling step.

Add an unit test. Before this change, the test fails at the point where
the test tries to read the CNAME packet and sees an incorrect type.